Fix typings not exporting an instance
